pyproject.toml
[project]
name = "llmwiki"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "openai>=1.0",
    "python-dotenv>=1.0",
    "rich>=15.0.0",
    "fastapi>=0.111",
    "uvicorn[standard]>=0.29",
]
